# frontend/src/api-generated/openapi.json
{"openapi":"3.0.0","info":{"title":"Momocraft Api","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"OAuthGoogleRequestDto":{"type":"object","properties":{"code":{"type":"string"},"redirectUrl":{"type":"string"}},"required":["code","redirectUrl"]},"UserProfileDto":{"type":"object","properties":{"userId":{"type":"string"},"avatarUrl":{"type":"string"},"email":{"type":"string"}},"required":["userId"]},"AuthedSessionDto":{"type":"object","properties":{"jwtToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserProfileDto"}},"required":["jwtToken","user"]},"OAuthRequestDto":{"type":"object","properties":{"code":{"type":"string"},"redirectUrl":{"type":"string"}},"required":["code","redirectUrl"]},"EmailAuthRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"ApiErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"},"error":{"type":"string"}},"required":["statusCode","message","error"]},"ItemCollectionDto":{"type":"object","properties":{"itemId":{"type":"string"},"state":{"type":"number"}},"required":["itemId","state"]},"UserCollectionListDto":{"type":"object","properties":{"collections":{"type":"array","items":{"$ref":"#/components/schemas/ItemCollectionDto"}}},"required":["collections"]},"UserFriendRequestDto":{"type":"object","properties":{"targetUserOrEmail":{"type":"string"},"comment":{"type":"string"},"requestMessage":{"type":"string"}},"required":["targetUserOrEmail","comment","requestMessage"]},"FriendUserDto":{"type":"object","properties":{"userId":{"type":"string"},"comment":{"type":"string"},"avatarUrl":{"type":"string"},"approvedAt":{"type":"number"}},"required":["userId","comment","approvedAt"]},"UserFriendCollectionDto":{"type":"object","properties":{"friend":{"$ref":"#/components/schemas/FriendUserDto"},"friendCollections":{"type":"array","items":{"$ref":"#/components/schemas/ItemCollectionDto"}}},"required":["friend","friendCollections"]},"FriendCollectionsDto":{"type":"object","properties":{"friendCollections":{"type":"array","items":{"$ref":"#/components/schemas/UserFriendCollectionDto"}}},"required":["friendCollections"]},"FriendListDto":{"type":"object","properties":{"friends":{"type":"array","items":{"$ref":"#/components/schemas/FriendUserDto"}}},"required":["friends"]}}},"paths":{"/auth/oauth/google":{"post":{"operationId":"AuthController_doGoogleOAuth","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OAuthGoogleRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthedSessionDto"}}}}}}},"/auth/oauth/discord":{"post":{"operationId":"AuthController_doDiscordOAuth","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OAuthRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthedSessionDto"}}}}}}},"/auth/email/signup":{"post":{"operationId":"AuthController_doEmailSignUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailAuthRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthedSessionDto"}}}}}}},"/auth/email/signin":{"post":{"operationId":"AuthController_doEmailSignIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailAuthRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthedSessionDto"}}}}}}},"/auth/jwt/refresh":{"post":{"operationId":"AuthController_doRefreshToken","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthedSessionDto"}}}}}}},"/auth/dummy/do-not-call":{"get":{"operationId":"AuthController_dummyDoNotCall","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}}}}},"/momo/user/{userId}/collections":{"get":{"operationId":"MomoUserController_getCollections","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCollectionListDto"}}}}}},"put":{"operationId":"MomoUserController_putCollections","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCollectionListDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCollectionListDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCollectionListDto"}}}}}}},"/momo/user/{userId}/friends":{"put":{"operationId":"MomoUserController_saveUserFriendRequest","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserFriendRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendUserDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendUserDto"}}}}}},"get":{"operationId":"MomoUserController_listFriends","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendListDto"}}}}}}},"/momo/user/{userId}/friendCollections":{"get":{"operationId":"MomoUserController_listFriendCollections","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendCollectionsDto"}}}}}}},"/momo/user/self":{"put":{"operationId":"MomoUserController_putSelfMeta","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileDto"}}}}}}}}}